+++ addon/mixins/flexberry-lookup-controller.js
@@ -219,6 +219,82 @@ export default Ember.Mixin.create(ReloadListMixin, {
     },
 
     /**
+      Handlers action from FlexberryLookup preview action.
+
+      @method actions.previewLookupValue
+      @param {Object} previewData Lookup parameters: { recordId, transitionRoute, transitionOptions, showInSeparateRoute, projection, modelName, controller }.
+    */
+    previewLookupValue(previewData) {
+      let options = Ember.$.extend(true, {
+        recordId: undefined,
+        transitionRoute: undefined,
+        transitionOptions: undefined,
+        showInSeparateRoute: undefined,
+        modelName: undefined,
+        controller: undefined,
+        projection: undefined
+      }, previewData);
+      let recordId = options.recordId;
+      let transitionRoute = options.transitionRoute;
+
+      if (options.showInSeparateRoute) {
+        let transitionOptions = options.transitionOptions || {};
+        this.transitionToRoute(transitionRoute, recordId, transitionOptions);
+      } else {
+        let routeName = options.controller || transitionRoute;
+        let modelName = options.modelName;
+
+        let controller = Ember.getOwner(this).lookup(`controller:${routeName}`);
+        if (Ember.isNone(controller)) {
+          throw new Error(`Controller with '${routeName}' name does not exist.`);
+        }
+
+        let route = Ember.getOwner(this).lookup(`route:${transitionRoute}`);
+        let projectionName = options.projection || (route ? route.get('modelProjection') : undefined);
+        if (Ember.isNone(projectionName)) {
+          throw new Error('\`previewFormProjection\` is undefined.');
+        }
+
+        let modelConstructor = this.store.modelFor(modelName);
+        let projection = Ember.get(modelConstructor, `projections.${projectionName}`);
+        if (!projection) {
+          throw new Error(
+            `No projection with '${projectionName}' name defined in '${modelName}' model.`);
+        }
+
+        controller.setProperties({
+          readonly: true,
+          routeName: routeName,
+          modelProjection: projection
+        });
+
+        let lookupController = this.get('lookupController');
+        lookupController.setProperties({
+          title: this.get('i18n').t('components.flexberry-lookup.preview-button-text'),
+          sizeClass: 'small preview-model',
+        });
+
+        let lookupSettings = this.get('lookupSettings');
+        this.send('showModalDialog', lookupSettings.template);
+
+        let loadingParams = {
+          view: lookupSettings.template,
+          outlet: 'modal-content'
+        };
+
+        this.send('showModalDialog', lookupSettings.loaderTemplate, null, loadingParams);
+
+        this.store.findRecord(modelName, recordId).then(data => {
+          this.send('removeModalDialog', loadingParams);
+          this.send('showModalDialog', transitionRoute, {
+            controller: controller,
+            model: data
+          }, loadingParams);
+        });
+      }
+    },
+
+    /**
       Update relation value at model.
 
       @method actions.updateLookupValue
@@ -311,7 +387,7 @@ export default Ember.Mixin.create(ReloadListMixin, {
       reloadData.saveTo);
 
     let modelConstructor = currentContext.store.modelFor(reloadData.relatedToType);
-    let projection = Ember.get(modelConstructor, 'projections')[reloadData.projectionName];
+    let projection = Ember.get(modelConstructor, `projections.${reloadData.projectionName}`);
     if (!projection) {
       throw new Error(
         `No projection with '${reloadData.projectionName}' name defined in '${reloadData.relatedToType}' model.`);