+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
 ### Added
+* `flexberry-lookup` component:
+    * Add preview button.
 * `flexberry-file` component:
     * Added the ability to open files instead of downloading.